rename: Make CreateSchedule directory
